Made changes for FireFox 3 support.
This patch adds support for FireFox 3 cookies file format, which is no longer
based on plain text files, but on sqlite databases.

The approach is to be as lean as possible on the requirements. Thus, there
is a double implementation, one for DBI/DBD::SQLite (if available), one
leaning on the command-line sqlite/sqlite3 program. Keeping this is
obviously a matter of taste.

There are also tests and an example cookies file, taken as copy of the
textual cookie file already present in the distribution.

git-svn-id: http://brian-d-foy.svn.sourceforge.net/svnroot/brian-d-foy/trunk@2590 d9f60a02-fc28-0410-9752-bc0d6c4f57a6